Morpho & Aave StavaV2 Rate Providers
Please review the new Static-A-token implementation from Aave which just went live. In my findings the primary change is the rewards rescuer will be the ALC Multisig so the Balancer ecosystem will need to coordinate with the Aave ecosystem on any pass through LM incentives being used on the lending market to re-distribute them to the pool participants. This is not a major change from the previous design.

MetaMorpho vaults are newer to me, however they seem to have far less upgradeable components. Overall the main concerns may arise around the curator or owner of the market being able to upgrade the fees which affect the convertToAssets accrual function to some degree based on how much fee is taken. The only cooldown noticeable is the time constraint for a curator to add new strategies to their vault, so LPs can decide to exit before they go into affect.

Multisig management on these vaults is relatively standard to sub-standard. 3/7 and or 2/6 based on my reviews.

In summary this includes:
Aave v3 Mainnet: - USDC - USDT - wETH
Aave Mainnet Lido Instance: - wETH
Aave Gnosis v3: - wETH - GNO - USDC.e